+++ tests/e2e/game/game-crud.test.js
@@ -142,7 +142,7 @@ describe("A - Game CRUD [Create / Read / Update / Delete]", () => {
                 done();
             });
     });
-    it(`🎲 Can't get an unknown game (GET /game/:id)`, done => {
+    it(`❓  Can't get an unknown game (GET /game/:id)`, done => {
         chai.request(server)
             // eslint-disable-next-line new-cap
             .get(`/games/${mongoose.Types.ObjectId()}`)
@@ -159,6 +159,42 @@ describe("A - Game CRUD [Create / Read / Update / Delete]", () => {
                 expect(res).to.have.status(200);
                 const game = res.body;
                 expect(game._id).to.be.equal(firstGame._id);
+                done();
+            });
+    });
+    it(`🔒 Can't update a game without basic authentication (PATCH /games/:id)`, done => {
+        chai.request(server)
+            .patch(`/games/${firstGame._id}`)
+            .send({ status: "playing" })
+            .end((err, res) => {
+                expect(res).to.have.status(401);
+                expect(res.body.type).to.equal("UNAUTHORIZED");
+                done();
+            });
+    });
+    it(`❓  Can't update an unknown game (PATCH /games/:id)`, done => {
+        chai.request(server)
+            // eslint-disable-next-line new-cap
+            .patch(`/games/${mongoose.Types.ObjectId()}`)
+            .auth(Config.app.basicAuth.username, Config.app.basicAuth.password)
+            .send({ status: "playing" })
+            .end((err, res) => {
+                expect(res).to.have.status(404);
+                expect(res.body.type).to.equal("NOT_FOUND");
+                done();
+            });
+    });
+    it(`🎲 Update the second game created (PATCH /games/:id)`, done => {
+        chai.request(server)
+            .patch(`/games/${firstGame._id}`)
+            .auth(Config.app.basicAuth.username, Config.app.basicAuth.password)
+            .send({ status: "playing" })
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                const game = res.body;
+                expect(game._id).to.equal(firstGame._id);
+                firstGame = game;
+                expect(firstGame.status).to.equal("playing");
                 done();
             });
     });